Use functions instead of a class for artifact authentication
The logic involves no state, so a class isn't necessary. And using
a class encourages dependency injection of a singleton, making the
code more involved than it needs to be.
