+++ libs/auth/src/index.ts
@@ -1,4 +1,4 @@
-export * from './artifact_authenticator';
+export * from './artifact_authentication';
 export type { CardStatus } from './card';
 export * from './config';
 export * from './dipped_smart_card_auth_api';
